fix: refactor layouts and routes structure
