+++ src/app/auth/login/page.tsx
@@ -1,7 +1,12 @@
-"use client";
+'use client'
 
-import LoginForm from "@/features/auth/_components/login-form";
+import LoginForm from '@/features/auth/_components/login-form'
+import AuthLayout from '../auth_layout/auth-layout'
 
 export default function Page() {
-  return <LoginForm />;
+  return (
+    <AuthLayout>
+      <LoginForm />
+    </AuthLayout>
+  )
 }
