+++ src/app/auth/register/page.tsx
@@ -1,7 +1,12 @@
 "use client";
 
 import SignUpForm from "@/features/auth/_components/sign-up-form";
+import AuthLayout from "../auth_layout/auth-layout";
 
 export default function Page() {
-  return <SignUpForm />;
+  return(
+    <AuthLayout>
+      <SignUpForm />;
+    </AuthLayout>
+  )
 }
