+++ src/app/auth/reset/page.tsx
@@ -1,7 +1,12 @@
-"use client";
+'use client'
 
-import ResetPasswordForm from "@/features/auth/_components/reset-password-form";
+import ResetPasswordForm from '@/features/auth/_components/reset-password-form'
+import AuthLayout from '../auth_layout/auth-layout'
 
 export default function Page() {
-  return <ResetPasswordForm />;
+  return (
+    <AuthLayout>
+      <ResetPasswordForm />
+    </AuthLayout>
+  )
 }
