--- src/app/courses/layout.tsx
@@ -1,9 +1,0 @@
-import SharedLayout from "@/app/shared/layout";
-
-export default function CourseLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return <SharedLayout>{children}</SharedLayout>;
-}
