--- src/app/dashboard/layout.tsx
@@ -1,9 +1,0 @@
-import SharedLayout from "../shared/layout";
-
-export default function DashboardLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return <SharedLayout>{children}</SharedLayout>;
-}
