+++ src/app/layout.tsx
@@ -1,25 +1,21 @@
-"use client";
+'use client'
 
-import React from "react";
-import { Comfortaa, Open_Sans } from "next/font/google";
-import "../styles/globals.css";
-import { Providers } from "./provider";
-import { ViewModeProvider } from "@/context/view-mode-context";
+import React from 'react'
+import { Comfortaa, Open_Sans } from 'next/font/google'
+import '../styles/globals.css'
+import { Providers } from './provider'
+import { ViewModeProvider } from '@/context/view-mode-context'
 
 const comfortaa = Comfortaa({
-  subsets: ["latin"],
-  weight: ["300", "400", "700"],
-});
+  subsets: ['latin'],
+  weight: ['300', '400', '700'],
+})
 const openSans = Open_Sans({
-  subsets: ["latin"],
-  weight: ["300", "400", "600", "700", "800"],
-});
+  subsets: ['latin'],
+  weight: ['300', '400', '600', '700', '800'],
+})
 
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
+export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="en" className={`${comfortaa.className} ${openSans.className}`}>
       <body className={comfortaa.className}>
@@ -28,5 +24,5 @@ export default function RootLayout({
         </Providers>
       </body>
     </html>
-  );
+  )
 }
