+++ src/app/page.tsx
@@ -1,6 +1,34 @@
 // app/page.tsx
-import { redirect } from "next/navigation";
 
-export default function Home() {
-  redirect("/dashboard");
+'use client'
+
+import { useEffect, useState } from 'react'
+import { useRouter } from 'next/navigation'
+import LoginPage from './auth/login/page'
+
+export default function HomePage() {
+  const [authChecked, setAuthChecked] = useState(false)
+  const [isAuth, setIsAuth] = useState(false)
+  const router = useRouter()
+
+  useEffect(() => {
+    const auth = checkAuthentication()
+    if (auth) {
+      setIsAuth(true)
+      router.replace('/dashboard')
+    } else {
+      setIsAuth(false)
+    }
+    setAuthChecked(true)
+  }, [router])
+
+  if (!authChecked) {
+    return null
+  }
+
+  return <LoginPage />
+}
+
+function checkAuthentication(): boolean {
+  return true
 }
